+++ SleepModel/HEMDeviceService.h
@@ -91,6 +91,8 @@ typedef void(^HEMDeviceResetHandler)(NSError* _Nullable error);
 - (BOOL)hasHardwareUpgradeForSense;
 - (BOOL)isPillFirmwareUpdateAvailable;
 - (SENSenseHardware)savedHardwareVersion;
+- (BOOL)savedHardwareVersionIsVoice;
+- (BOOL)hasSavedHardwareVersion;
 
 /**
  * @return YES if we should show pill information to the users, NO otherwise